requirements.txt
discord.py==2.3.2
aiohttp==3.9.5
redis[asyncio]==5.0.8
python-dotenv==1.0.1
